两个线程接收消息

+++ central_node/imu.h
@@ -59,7 +59,7 @@ typedef struct {
     float N_az;
 
     float ADC;
-    float GPIO;
+    int GPIO;
     
 }ImuData;
